+++ src/capacitor.rs
@@ -4,12 +4,12 @@ use circuit::{NodeId};
 pub struct Capacitor {
     pub a: NodeId,
     pub b: NodeId,
-    pub value: f32, // Farads
+    pub value: f64, // Farads
 }
 
 impl Capacitor {
 
-    pub fn linearize(&self, v: f32, t: f32) -> (f32, f32) {
+    pub fn linearize(&self, v: f64, t: f64) -> (f64, f64) {
         let g_eq = self.value / t;
         let i_eq = g_eq * v;
         (g_eq, i_eq)