+++ tests/common/mod.rs
@@ -1,6 +1,6 @@
 #[allow(dead_code)]
-pub fn assert_nearly(x: f32, expected: f32) {
-    const EPSILON: f32 = 1e-5;
+pub fn assert_nearly(x: f64, expected: f64) {
+    const EPSILON: f64 = 1e-5;
     let delta = (x - expected).abs();
     assert!( delta < EPSILON, "{} isn't approximately {}", x, expected);
 }